backend/requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
click==8.3.1
email-validator==2.3.0
fastapi==0.132.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
SQLAlchemy==2.0.46
starlette==0.52.1
uvicorn==0.41.0
